+++ Source/Singulink.Cryptography.PasswordHasher/PasswordHasher.cs
@@ -2,22 +2,19 @@ using System.Diagnostics;
 using System.Globalization;
 using System.Security.Cryptography;
 using System.Text;
+using Microsoft.Extensions.Options;
 
 namespace Singulink.Cryptography;
 
 /// <summary>
 /// Provides upgradable password hashing functionality. All operations are thread-safe.
 /// </summary>
-public sealed class PasswordHasher
+public sealed class PasswordHasher : IPasswordHasher
 {
     private const char Separator = ' ';
 
-#if !NET
-    private static readonly RNGCryptoServiceProvider CryptoRandom = new();
-#endif
-
-    private readonly Dictionary<string, PasswordHashAlgorithm> _algorithmLookup = new();
-    private readonly Dictionary<int, HashEncryptionParameters> _encryptionLookup = new();
+    private readonly Dictionary<string, PasswordHashAlgorithm> _algorithmLookup = [];
+    private readonly Dictionary<int, HashEncryptionParameters> _encryptionLookup = [];
 
     /// <summary>
     /// Gets the main password hashing algorithm.
@@ -57,74 +54,142 @@ public sealed class PasswordHasher
     /// <summary>
     /// Initializes a new instance of the <see cref="PasswordHasher"/> class.
     /// </summary>
-    /// <param name="algorithm">The password hashing algorithm.</param>
-    /// <param name="iterations">The number of hashing iterations to perform.</param>
-    /// <param name="options">Any additional options that should be applied to the password hashesr.</param>
-    public PasswordHasher(PasswordHashAlgorithm algorithm, int iterations, PasswordHasherOptions? options = null)
+    /// <param name="options">Any additional options that should be applied to the password hasher.</param>
+    public PasswordHasher(PasswordHasherOptions options)
     {
-#pragma warning disable CS0618 // Type or member is obsolete
+        Algorithm = options.Algorithm;
+        _algorithmLookup.Add(Algorithm.Id, Algorithm);
+        Iterations = options.Iterations;
 
-        if (algorithm == PasswordHashAlgorithm.SHA1)
-            throw new ArgumentException("SHA1 is not considered safe and is only supported as a legacy algorithm.", nameof(algorithm));
+        SaltSize = options.SaltSize;
+        Normalize = options.Normalize;
 
-#pragma warning restore CS0618 // Type or member is obsolete
-
-        if (iterations <= 0)
-            throw new ArgumentOutOfRangeException(nameof(iterations));
-
-        Algorithm = algorithm;
-        _algorithmLookup.Add(algorithm.Id, algorithm);
-        Iterations = iterations;
-
-        if (options == null)
+        if (options.EncryptionParameters is { } ep)
         {
-            SaltSize = PasswordHasherOptions.DefaultSaltSize;
-            Normalize = PasswordHasherOptions.DefaultNormalize;
+            EncryptionParameters = ep;
+            _encryptionLookup.Add(ep.Id, ep);
         }
-        else
+
+        foreach (var a in options.LegacyHashAlgorithms)
         {
-            SaltSize = options.SaltSize;
-            Normalize = options.Normalize;
+            if (!_algorithmLookup.TryAdd(a.Id, a))
+                throw new ArgumentException("Hash algorithms must all have unique IDs.", nameof(options));
+        }
 
-            if (options.EncryptionParameters is { } ep)
-            {
-                EncryptionParameters = ep;
-                _encryptionLookup.Add(ep.Id, ep);
-            }
-
-            foreach (var a in options.LegacyHashAlgorithms)
-            {
-                if (!_algorithmLookup.TryAdd(a.Id, a))
-                    throw new ArgumentException("Hash algorithms must all have unique IDs.", nameof(options));
-            }
-
-            foreach (var p in options.LegacyEncryptionParameters)
-            {
-                if (!_encryptionLookup.TryAdd(p.Id, p))
-                    throw new ArgumentException("Encryption parameters must all have unique IDs.", nameof(options));
-            }
+        foreach (var p in options.LegacyEncryptionParameters)
+        {
+            if (!_encryptionLookup.TryAdd(p.Id, p))
+                throw new ArgumentException("Encryption parameters must all have unique IDs.", nameof(options));
         }
     }
 
     /// <summary>
-    /// Hashes the specified password.
+    /// Initializes a new instance of the <see cref="PasswordHasher"/> class with the specified options.
     /// </summary>
-    /// <param name="password">The password to hash.</param>
-    /// <returns>A string containing the normalization setting, master key ID, algorith ID, number of iterations, salt value and password hash.</returns>
-    /// <exception cref="ArgumentException">
-    /// If normalization is enabled then password contained invalid Unicode characters or disallowed characters (i.e. control characters).
-    /// </exception>
+    public PasswordHasher(IOptions<PasswordHasherOptions> options) : this(options.Value) { }
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="PasswordHasher"/> class with the specified algorithm and iteration count and optional builder function that
+    /// configure additional options.
+    /// </summary>
+    /// <param name="algorithm">The main password hashing algorithm.</param>
+    /// <param name="iterations">The number of hashing iterations to perform.</param>
+    /// <param name="optionsBuilder">Optional builder function that configures additional options.</param>
+    public PasswordHasher(PasswordHashAlgorithm algorithm, int iterations, Action<PasswordHasherOptions>? optionsBuilder = null)
+        : this(BuildOptions(algorithm, iterations, optionsBuilder)) { }
+
+    private static PasswordHasherOptions BuildOptions(PasswordHashAlgorithm algorithm, int iterations, Action<PasswordHasherOptions>? optionsBuilder)
+    {
+        var options = new PasswordHasherOptions(algorithm, iterations);
+        optionsBuilder?.Invoke(options);
+        return options;
+    }
+
+    /// <inheritdoc cref="IPasswordHasher.Hash(string)"/>
     public string Hash(string password)
     {
         return GetPreamble(Normalize) + HashWithoutPreamble(GetPasswordBytes(password, Normalize), Iterations);
     }
 
-    /// <summary>
-    /// Validates a password against a hash string.
-    /// </summary>
-    /// <param name="hash">The hash string representing the password.</param>
-    /// <param name="password">The password to validate.</param>
-    /// <returns>True if the password is correct, otherwise false.</returns>
+    /// <inheritdoc cref="IPasswordHasher.Rehash(string)"/>
+    public string Rehash(string password)
+    {
+        byte[] passwordBytes;
+        bool normalized;
+
+        try
+        {
+            passwordBytes = GetPasswordBytes(password, Normalize);
+            normalized = Normalize;
+        }
+        catch (ArgumentException) when (Normalize)
+        {
+            passwordBytes = GetPasswordBytes(password, false);
+            normalized = false;
+        }
+
+        return GetPreamble(normalized) + HashWithoutPreamble(passwordBytes, Iterations);
+    }
+
+    /// <inheritdoc cref="IPasswordHasher.RequiresRehash(string, string)"/>
+    public bool RequiresRehash(string hash, string password)
+    {
+        (bool normalized, var encryptionParameters, string[] hashParams, _) = GetHashParts(hash);
+
+        if ((!normalized && Normalize && CanNormalize(password)) || (normalized && !Normalize))
+            return true;
+
+        if (encryptionParameters != EncryptionParameters || hashParams.Length > 1)
+            return true;
+
+        var info = HashAlgorithmInfo.Parse(hashParams[0], _algorithmLookup);
+        return info.Algorithm != Algorithm || info.Iterations != Iterations;
+    }
+
+    /// <inheritdoc cref="IPasswordHasher.RequiresUpdate(string)"/>
+    public bool RequiresUpdate(string hash)
+    {
+        (_, var encryptionParameters, string[] hashParams, _) = GetHashParts(hash);
+
+        if (encryptionParameters != EncryptionParameters)
+            return true;
+
+        int actualIterations = GetMainHashIterations(hashParams);
+        return actualIterations < Iterations;
+    }
+
+    /// <inheritdoc cref="IPasswordHasher.Update(string)"/>
+    public string Update(string hash)
+    {
+        (bool normalized, var encryptionParameters, string[] hashParams, byte[] hashBytes) = GetHashParts(hash);
+
+        int extraIterations = Iterations - GetMainHashIterations(hashParams);
+
+        if (extraIterations <= 0 && encryptionParameters == EncryptionParameters)
+            return hash;
+
+        if (encryptionParameters != null)
+            hashBytes = encryptionParameters.Algorithm.Decrypt(encryptionParameters.Key, hashBytes);
+
+        // The new hash will contain the new encryption parameters so just chain hashes if we need extra iterations, otherwise just encrypt the hash bytes
+        // with the new key. Preamble needs to be regenerated in each case based on previous normalization settings as updating the hash does not change
+        // the normalization.
+
+        if (extraIterations > 0)
+        {
+            string newHashEnding = HashWithoutPreamble(hashBytes, extraIterations);
+            return $"{GetPreamble(normalized)}{string.Join(' ', hashParams)} {newHashEnding}";
+        }
+        else
+        {
+            if (EncryptionParameters != null)
+                hashBytes = EncryptionParameters.Algorithm.Encrypt(EncryptionParameters.Key, hashBytes);
+
+            return $"{GetPreamble(normalized)}{string.Join(' ', hashParams)} {Convert.ToBase64String(hashBytes)}";
+        }
+    }
+
+    /// <inheritdoc cref="IPasswordHasher.Verify(string, string)"/>
     public bool Verify(string hash, string password)
     {
         if (password.Length == 0)
@@ -155,115 +220,12 @@ public sealed class PasswordHasher
         return SlowEquals(passwordHashBytes, hashBytes);
     }
 
-    /// <summary>
-    /// Returns a value indicating whether a hash should be regenerated from the known password. Returns <see langword="true"/> if the hash contains
-    /// chained hashes, the main algorithm / number of iterations does not match, the main encryption parameters do not match, or normalization settings do
-    /// not match.
-    /// </summary>
-    public bool RequiresRehash(string hash, string password)
-    {
-        (bool normalized, var encryptionParameters, string[] hashParams, _) = GetHashParts(hash);
-
-        if ((!normalized && Normalize && CanNormalize(password)) || (normalized && !Normalize))
-            return true;
-
-        if (encryptionParameters != EncryptionParameters || hashParams.Length > 1)
-            return true;
-
-        var info = HashAlgorithmInfo.Parse(hashParams[0], _algorithmLookup);
-        return info.Algorithm != Algorithm || info.Iterations != Iterations;
-    }
-
-    /// <summary>
-    /// Safely rehashes an existing password by falling back to previous normalization settings if normalization fails with current settings.
-    /// </summary>
-    /// <param name="password">The password to hash.</param>
-    /// <returns>A string containing the normalization setting, master key ID, algorith ID, number of iterations, salt value and password hash.</returns>
-    public string Rehash(string password)
-    {
-        byte[] passwordBytes;
-        bool normalized;
-
-        try
-        {
-            passwordBytes = GetPasswordBytes(password, Normalize);
-            normalized = Normalize;
-        }
-        catch (ArgumentException) when (Normalize)
-        {
-            passwordBytes = GetPasswordBytes(password, false);
-            normalized = false;
-        }
-
-        return GetPreamble(normalized) + HashWithoutPreamble(passwordBytes, Iterations);
-    }
-
-    /// <summary>
-    /// Returns a value indicating whether the hash needs to be updated. Returns <see langword="true"/> if the hash chain needs to be updated so that it
-    /// utilizes the main algorithm and total required number of iterations. Also returns true if the main encryption parameters do not match.
-    /// </summary>
-    public bool RequiresUpdate(string hash)
-    {
-        (_, var encryptionParameters, string[] hashParams, _) = GetHashParts(hash);
-
-        if (encryptionParameters != EncryptionParameters)
-            return true;
-
-        int actualIterations = GetMainHashIterations(hashParams);
-        return actualIterations < Iterations;
-    }
-
-    /// <summary>
-    /// Gets an updated hash that uses the main encryption parameters and main hash algorithm with the total number of required iterations without knowing
-    /// the password, or returns <see langword="null"/> if hash does not require an update.
-    /// </summary>
-    /// <remarks>
-    /// <para>Changing hash algorithms or adding iterations without knowing the password is achieved by hash chaining. If the hash algorithm or number of
-    /// iterations has changed then the resulting hash will return <see langword="true"/> when passed into the <see cref="RequiresRehash(string, string)"/> method, which
-    /// should be tested on successful user login so that a new hash without chaining can be generated with the <see cref="Rehash(string)"/>
-    /// method.</para>
-    /// </remarks>
-    public string? Update(string hash)
-    {
-        (bool normalized, var encryptionParameters, string[] hashParams, byte[] hashBytes) = GetHashParts(hash);
-
-        int extraIterations = Iterations - GetMainHashIterations(hashParams);
-
-        if (extraIterations <= 0 && encryptionParameters == EncryptionParameters)
-            return null;
-
-        if (encryptionParameters != null)
-            hashBytes = encryptionParameters.Algorithm.Decrypt(encryptionParameters.Key, hashBytes);
-
-        // The new hash will contain the new encryption parameters so just chain hashes if we need extra iterations, otherwise just encrypt the hash bytes
-        // with the new key. Preamble needs to be regenerated in each case based on previous normalization settings as updating the hash does not change
-        // the normalization.
-
-        if (extraIterations > 0)
-        {
-            string newHashEnding = HashWithoutPreamble(hashBytes, extraIterations);
-            return $"{GetPreamble(normalized)}{string.Join(' ', hashParams)} {newHashEnding}";
-        }
-        else
-        {
-            if (EncryptionParameters != null)
-                hashBytes = EncryptionParameters.Algorithm.Encrypt(EncryptionParameters.Key, hashBytes);
-
-            return $"{GetPreamble(normalized)}{string.Join(' ', hashParams)} {Convert.ToBase64String(hashBytes)}";
-        }
-    }
-
     private string HashWithoutPreamble(byte[] data, int iterations)
     {
         Debug.Assert(data.Length > 0, "data cannot be empty");
 
         byte[] salt = new byte[SaltSize];
-
-#if NET
         RandomNumberGenerator.Fill(salt);
-#else
-        CryptoRandom.GetBytes(salt);
-#endif
 
         byte[] hashBytes = Algorithm.Hash(data, salt, iterations);
 
@@ -394,24 +356,17 @@ public sealed class PasswordHasher
         return diff == 0;
     }
 
-    private static Exception GetHashStringFormatException(string message) => new FormatException("Hash string was in an invalid format: " + message);
+    private static FormatException GetHashStringFormatException(string message) => new("Hash string was in an invalid format: " + message);
 
-    private struct HashAlgorithmInfo
+    private readonly struct HashAlgorithmInfo(PasswordHashAlgorithm algorithm, int iterations, byte[] salt)
     {
         private const char Separator = ':';
 
-        public PasswordHashAlgorithm Algorithm { get; }
+        public PasswordHashAlgorithm Algorithm => algorithm;
 
-        public int Iterations { get; }
+        public int Iterations => iterations;
 
-        public byte[] Salt { get; }
-
-        public HashAlgorithmInfo(PasswordHashAlgorithm algorithm, int iterations, byte[] salt)
-        {
-            Algorithm = algorithm;
-            Iterations = iterations;
-            Salt = salt;
-        }
+        public byte[] Salt => salt;
 
         public static HashAlgorithmInfo Parse(string s, Dictionary<string, PasswordHashAlgorithm> algorithms)
         {